deleting unused files and adding a c3po namespace
As of this change I can no longer render the cornell box but that is only temporary because I plan to have materials and triangle meshes working soon. If anyone other than me is looking at this repo currently just wait a couple days and it should be working again.